projeto_experimental4/requirements.txt
asgiref==3.8.1
asttokens==2.4.1
colorama==0.4.6
decorator==5.1.1
Django==4.1.13
djongo==1.3.6
dnslib==0.9.25
dnspython==2.6.1
exceptiongroup==1.2.2
executing==2.1.0
ipdb==0.13.13
ipython==8.27.0
jedi==0.19.1
matplotlib-inline==0.1.7
parso==0.8.4
pexpect==4.9.0
pillow==11.0.0
prompt_toolkit==3.0.48
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.18.0
pymongo==4.10.1
python-dateutil==2.9.0.post0
PythonDNS==0.1
pytz==2024.2
six==1.16.0
sqlparse==0.2.4
stack-data==0.6.3
tomli==2.0.2
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
wcwidth==0.2.13
